Inject `lookup` into dicts as entry function
It is the simple way to improve encapsulation.

It is still difficult to konw which type/const/fn are required
when implementing a new dict.
Moreover, the injection makes the coding style worse -- a common
struct comes from super, and a common function comes from macro.

A trait can fix it.